Update Program.cs
Added folder-processing support and length check to avoid renaming to longer than MAX_PATH and rendering items inaccessible.

NB: The /f switch was repurposed, so any scripts that used it should be updated.

+++ source/Program.cs
@@ -10,12 +10,17 @@ namespace RenameRegex
     using System.IO;
     using System.Reflection;
     using System.Text.RegularExpressions;
+    using System.Collections.Generic;
 
     /// <summary>
     /// RenameRegex command line program
     /// </summary>
     public static class Program
     {
+        const int MAX_PATH = 260;
+        public const int incFiles=1;
+        public const int incDirs=2;
+
         /// <summary>
         /// Main command line
         /// </summary>
@@ -31,6 +36,7 @@ namespace RenameRegex
             bool pretend;
             bool force;
             bool preserveExt;
+            int  include;
 
             if (!GetArguments(
                     args,
@@ -40,21 +46,32 @@ namespace RenameRegex
                     out pretend,
                     out recursive,
                     out force,
-                    out preserveExt))
+                    out preserveExt,
+                    out include))
             {
                 Usage();
                 return 1;
             }
 
-            // enumerate all files
+            // enumerate all files and directories
+            List<string> allItems = new List<string>();
+
             string[] files = Directory.GetFiles(
                 System.Environment.CurrentDirectory,
                 fileMatch,
                 recursive ? SearchOption.AllDirectories : SearchOption.TopDirectoryOnly);
 
-            if (files.Length == 0)
+            string[] dirs = Directory.GetDirectories(
+                System.Environment.CurrentDirectory,
+                fileMatch,
+                recursive ? SearchOption.AllDirectories : SearchOption.TopDirectoryOnly);
+
+            if ((include==0) || ((include&incFiles)!=0)) allItems.AddRange(files);
+            if ((include==0) || ((include&incDirs)!=0))  allItems.AddRange(dirs);
+    
+            if (allItems.Count == 0)
             {
-                Console.WriteLine(@"No files match!");
+                Console.WriteLine(@"No files or directories match!");
                 return 1;
             }
 
@@ -63,8 +80,13 @@ namespace RenameRegex
             //
             // loop through each file, renaming via a regex
             //
-            foreach (string fullFile in files)
+            foreach (string fullFile in allItems)
             {
+                if (fullFile.Length > MAX_PATH || Path.GetDirectoryName(fullFile).Length > MAX_PATH-12) {
+                    Console.WriteLine(@"""{0}"" cannot be accessed; too long.", fullFile);
+                    continue;
+                }
+
                 // split into filename, extension and path
                 string fileName = Path.GetFileNameWithoutExtension(fullFile);
                 string fileExt  = Path.GetExtension(fullFile);
@@ -78,7 +100,7 @@ namespace RenameRegex
                 }
 
                 // rename via a regex
-                string fileNameAfter = Regex.Replace(fileName, nameSearch, nameReplace, RegexOptions.IgnoreCase);
+                string fileNameAfter = Regex.Replace(fileName, nameSearch, nameReplace);
 
                 if (preserveExt)
                 {
@@ -99,7 +121,7 @@ namespace RenameRegex
                         (fileDir + @"\" + fileName).Replace(System.Environment.CurrentDirectory + @"\", String.Empty);
 
                     Console.WriteLine(
-                        @"{0} -> {1}{2}{3}",
+                        @"{0} -> " + Environment.NewLine + "{1}{2}{3}" + Environment.NewLine,
                         fileNameToShow,
                         fileNameAfter,
                         pretendModeNotification,
@@ -117,7 +139,12 @@ namespace RenameRegex
                             File.Delete(fileNameAfter);
                         }
 
-                        File.Move(fileDir + @"\" + fileName, fileDir + @"\" + fileNameAfter);
+                        if (File.Exists(fileName))
+													File.Move(fileDir + @"\" + fileName, fileDir + @"\" + fileNameAfter);
+												else if (Directory.Exists(fileName))
+													Directory.Move(fileDir + @"\" + fileName, fileDir + @"\" + fileNameAfter);
+												else
+													Console.WriteLine(@"Could not rename {0}.",  fileName);
                     }
                     catch (IOException)
                     {
@@ -140,6 +167,7 @@ namespace RenameRegex
         /// <param name="recursive">Whether or not to recursively look in directories</param>
         /// <param name="force">Whether or not to force overwrites</param>
         /// <param name="preserveExt">Whether or not to preserve file extensions</param>
+        /// <param name="include">Whether to include directories, files or both</param>
         /// <returns>True if argument parsing was successful</returns>
         private static bool GetArguments(
             string[] args,
@@ -149,7 +177,8 @@ namespace RenameRegex
             out bool pretend,
             out bool recursive,
             out bool force,
-            out bool preserveExt)
+            out bool preserveExt,
+            out int  include)
         {
             // defaults
             fileMatch   = String.Empty;
@@ -162,6 +191,7 @@ namespace RenameRegex
             recursive   = false;
             force       = false;
             preserveExt = false;
+            include     = 0;
 
             // check for all arguments
             if (args == null || args.Length < 3)
@@ -188,13 +218,21 @@ namespace RenameRegex
                 {
                     recursive = true;
                 }
-                else if (args[i].Equals("/f", StringComparison.OrdinalIgnoreCase))
+                else if (args[i].Equals("/q", StringComparison.OrdinalIgnoreCase) || args[i].Equals("/y", StringComparison.OrdinalIgnoreCase))
                 {
                     force = true;
                 }
                 else if (args[i].Equals("/e", StringComparison.OrdinalIgnoreCase))
                 {
                     preserveExt = true;
+                }
+                else if (args[i].Equals("/f", StringComparison.OrdinalIgnoreCase))
+                {
+                    include |= incFiles;
+                }
+                else if (args[i].Equals("/d", StringComparison.OrdinalIgnoreCase))
+                {
+                    include |= incDirs;
                 }
                 else
                 {
@@ -236,8 +274,11 @@ namespace RenameRegex
             Console.WriteLine(@"Usage: RR.exe file-match search replace [/p] [/r] [/f] [/e]");
             Console.WriteLine(@"        /p: pretend (show what will be renamed)");
             Console.WriteLine(@"        /r: recursive");
-            Console.WriteLine(@"        /f: force overwrite if the file already exists");
+            Console.WriteLine(@"     /q|/y: force overwrite if the file already exists");
             Console.WriteLine(@"        /e: preserve file extensions");
+            Console.WriteLine(@"        /f: include only files");
+            Console.WriteLine(@"        /d: include only directories");
+            Console.WriteLine(@"            default is to include files and folders");
             return;
         }
     }
